pyproject.toml
[project]
name = "llm-prj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "flagembedding>=1.2.9",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.7",
    "langchain-milvus>=0.2.1",
    "langdetect>=1.0.9",
    "langgraph>=1.0.0",
    "pymilvus>=2.6.2",
    "pymupdf>=1.26.4",
    "transformers>=4.40.2",
    "elasticsearch>=9.1.1",
    "fastapi>=0.121.0",
    "uvicorn>=0.38.0",
    "pymysql>=1.1.2",
    "pydantic[email]>=2.12.3",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "passlib>=1.7.4",
    "bcrypt==4.3.0",
]
# torch는 GPU 버전 사용으로 삭제했다가 다시 설치를 해야함 -> 예시: [uv pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126]